Add flipper config
