+++ src/components/common/ManageToken/index.js
@@ -1,0 +1,4 @@
+// @flow
+export { default as TokenRow } from './TokenRow'
+export { default as TokensHeader } from './TokensHeader'
+export * from './helpers'
